refactor: move non-gui logic from notebook module to common module.
