+++ penelope/notebook/word_trends/__init__.py
@@ -2,6 +2,5 @@
 from .displayers import DEFAULT_WORD_TREND_DISPLAYERS, BarDisplayer, ITrendDisplayer, LineDisplayer, TableDisplayer
 from .gof_and_trends_gui import GofTrendsGUI
 from .gofs_gui import GoFsGUI
-from .interface import BundleTrendsService, TrendsComputeOpts, TrendsService
 from .trends_gui import CoOccurrenceTrendsGUI, TrendsBaseGUI, TrendsGUI
 from .trends_with_picks_gui import TokensSelector, TrendsWithPickTokensGUI
